frontend: move all enums to /types/enums
