+++ frontend/types/enums/chromie-time.ts
