+++ frontend/types/enums/faction.ts
