+++ frontend/types/enums/instance-type.ts
@@ -1,0 +1,6 @@
+export enum InstanceType {
+    Dungeon = 1,
+    Raid = 2,
+    Pvp = 3,
+    Scenario = 5,
+}
