+++ frontend/types/enums/inventory-slot.ts
@@ -1,0 +1,26 @@
+export enum InventorySlot {
+    Ammo,
+    Head,
+    Neck,
+    Shoulders,
+    Shirt,
+    Chest,
+    Waist,
+    Legs,
+    Feet,
+    Wrist,
+    Hands,
+    Ring1,
+    Ring2,
+    Trinket1,
+    Trinket2,
+    Back,
+    MainHand,
+    OffHand,
+    Ranged,
+    Tabard,
+    Bag1,
+    Bag2,
+    Bag3,
+    Bag4,
+}
