+++ frontend/types/enums/playable-class.ts
