+++ frontend/utils/get-character-group-func.ts
@@ -1,8 +1,8 @@
 import { get } from 'svelte/store'
 
-import {Faction} from '@/data/faction'
 import { data as userData } from '@/stores/user'
 import type { Character} from '@/types'
+import {Faction} from '@/types/enums'
 
 export default function getCharacterGroupFunc(): (char: Character) => string {
     // TODO hook this up to settings
